requirements.txt
selenium==4.15.2
pytest==7.4.3
webdriver-manager==4.0.1
rpa==1.48.0
pyautogui
xvfbwrapper
python-dotenv==1.0.0
pyvirtualdisplay
pytest-html